requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
APScheduler==3.9.1
async-timeout==4.0.2
attrs==22.1.0
Bootstrap-Flask==2.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
dnspython==2.0.0
dominate==2.5.2
email-validator==1.1.1
Flask==2.2.2
Flask-Bootstrap==3.3.7.1
flask-sock==0.5.2
Flask-WTF==0.14.3
frozenlist==1.3.1
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
meraki==1.27.0
multidict==6.0.2
python-dotenv==0.21.0
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
simple-websocket==0.8.1
six==1.16.0
Turbo-Flask==0.8.2
tzdata==2022.5
tzlocal==4.2
urllib3==1.26.12
visitor==0.1.3
Werkzeug==2.2.2
wsproto==1.2.0
WTForms==2.3.3
yarl==1.8.1
